+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+- Add support for implicit multiplication to `GrammarParser`.
+
 ## [3.1.0] - 2025-06-29
 
 ### Changed
@@ -107,7 +113,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Renamed the existing expression parser to `ShuntingYardParser`. This parser
   is now in maintenance mode and will not receive any new features.
-  `Parser` remains as a deprecated type alias to this legacy implemenation to
+  `Parser` remains as a deprecated type alias to this legacy implementation to
   maintain backwards compatibility.
 
 ### Fixed